Create melted
This is the code in which I 'melt' the scraped info in order to facilitate the creation of my actor node list and edge list. Each line of the dataset is an actor's appearance in the a top movie of the era. For example, if Marilyn Monroe was present in 10 movies, there will be 10 appearances of her in the dataset.